+++ nomad/raft_rpc.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/hashicorp/nomad/tlsutil"
+	"github.com/hashicorp/nomad/helper/tlsutil"
 )
 
 // RaftLayer implements the raft.StreamLayer interface,
